market-insight-agent/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
openai>=1.12.0
beautifulsoup4>=4.12.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
lxml>=5.1.0
tavily-python>=0.3.3
slowapi>=0.1.9
structlog>=24.1.0
